folders restructured and basic profile styling
