+++ client/src/App.js
@@ -10,10 +10,10 @@ import { authenticate } from "./utils/auth";
 // Components
 import Login from "./components/auth/Login.js";
 import Register from "./components/auth/Register.js";
-import DuckNavbar from "./components/Navbar.js";
+import DuckNavbar from "./components/navbar/Navbar.js";
 import ProtectedRoute from "./components/protected/ProtectedRoute";
 import UserProfile from "./components/profile/UserProfile";
-import Leaderboard from "./components/Leaderboard";
+import Leaderboard from "./components/leaderboard/Leaderboard";
 import Feed from "./components/feed/Feed";
 
 // CSS
