+++ client/src/components/profile/RecentPosts.css
@@ -1,0 +1,3 @@
+.recent-posts {
+    background-color: rgb(202, 202, 202);
+}
