--- client/src/components/profile/UserInfo.css
@@ -1,3 +1,0 @@
-.user-info {
-    text-align: center;
-}
