+++ client/src/components/profile/UserProfile.css
@@ -1,0 +1,6 @@
+.user-info {
+    padding-top: 20px;
+    margin-top: 15px;
+    text-align: center;
+    background-color: rgb(204, 204, 204);
+}
