+++ client/src/components/profile/containers/ProfileInfo.css
@@ -1,7 +1,9 @@
-.user-info {
-    text-align: center;
+.image-body img {
+    border: 5px solid rgb(241, 241, 241);
+    box-shadow: rgb(153, 153, 153) 0px 0px 2px 1px;
+    border-radius: 50%;
 }
 
-.profile-info {
-    background-color: rgb(204, 204, 204);
+.user-names h4 {
+    font-size: 50px;
 }
