+++ client/src/components/profile/containers/ProfileInfo.js
@@ -12,8 +12,8 @@ export default class ProfileInfo extends Component {
   }
 
   displayAdminStatus = () => {
-    if (!this.state.isAdmin) return " Non-Admin";
-    else return " Admin";
+    if (!this.state.isAdmin) return " This User is Non-Admin";
+    else return " This User is Admin";
   };
 
   render() {
@@ -26,27 +26,35 @@ export default class ProfileInfo extends Component {
     } = this.props.data;
     return (
       <Media className="profile-info">
-        <Media body>
-          <Media heading>
-            <img
-              src={sampleImage}
-              alt={"DuckImg"}
-              height={"75px"}
-              width={"100px"}
-            />
-          </Media>
-          <Media heading>
-            <Badge color="primary">
-              {firstName} {lastName}
-            </Badge>
-          </Media>
-          <Media heading>
-            <Badge color="success">Email:</Badge> {email}
-          </Media>
-          <Media heading>
-            <Badge color="danger">Admin Status: </Badge>
-            {this.displayAdminStatus()}
-          </Media>
+        <Media body className="image-body row">
+          <div className="col-6">
+            <Media heading>
+              <img
+                src={sampleImage}
+                alt={"DuckImg"}
+                height={"175px"}
+                width={"175px"}
+              />
+            </Media>
+          </div>
+          <div className="col user-names">
+            <Media heading>
+              <Badge color="primary">
+                {firstName} {lastName}
+              </Badge>
+            </Media>
+          </div>
+        </Media>
+        <Media body className="row">
+          <div className="col details">
+            <Media heading>
+              <Badge color="success">Email:</Badge> {email}
+            </Media>
+            <Media heading>
+              <Badge color="danger">Admin Status: </Badge>
+              {this.displayAdminStatus()}
+            </Media>
+          </div>
         </Media>
       </Media>
     );
